+++ src/routes/users/user/index.ts
@@ -8,6 +8,7 @@ import unblock from "./actions/unblock";
 import mute from "./actions/mute";
 import unmute from "./actions/unmute";
 import edit from "./actions/edit";
+import uploadAvatar from "./uploadAvatar";
 
 export default function (
     fastify: FastifyInstance,
@@ -17,6 +18,7 @@ export default function (
     fastify.register(profile);
     fastify.register(name);
     fastify.register(avatar);
+    fastify.register(uploadAvatar);
     fastify.register(threads);
     fastify.register(block);
     fastify.register(unblock);